+++ release_notes.md
@@ -1,3 +1,8 @@
+### 1.0.2 (unreleased)
+* Expose TapThreshold property to allow tweaking of tap detection.
+* Add option to detect taps using mix of translation and velocity.
+  Based on analysis performed by Ryan Melman.
+
 ### 1.0.1 (Released May 2015)
 
 * Fixed nuget packaging issue that didn't add a reference to this library.